+++ src/test/java/constant/AttendanceStatusTest.java
@@ -1,5 +1,6 @@
 package constant;
 
+import domain.AttendanceStatus;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import org.assertj.core.api.SoftAssertions;
